Update non-incident support workflow
I want to share with you a first iteration of a workflow for our support stewards. The whole idea behind this workflow is to start with the process of standardizing our support response. Historically, our support response has been dependent on the support steward and bias towards a reactive response we do not control properly. This simple workflow tried to battle that bias whereas it is bringing a common pattern so all of our support responses are cohesive and shared among our support stewards.

More details in this Slack message: https://2i2c.slack.com/archives/C028WU9PFBN/p1691789689379759