Note books related to responsibilities are reimbursible (#790)
I was going to buy and reimburse https://www.oreilly.com/library/view/the-staff-engineers/9781098118723/,
which I think is very helpful for my role in the team. However,
it isn't entirely clear if that is reimbursible (as it isn't
'training', and 'personal development' feels too broad?).

So explicitly list 'books' here. This isn't *any* book - the
filter in the latter part of the sentence constraints this well
still.