Require support tickets to be *initiated* by community reps
I'm working with Sarah in systematizing how we would
respond to a request like https://2i2c.freshdesk.com/a/tickets/1259.
Among many issues, one is that we don't actually know who the person
asking for this is or if they are authorized to ask for this change.

See https://github.com/2i2c-org/infrastructure/issues/3048 for
the process of how the linked airtable came to be.